requirements.txt

Flask==2.2.5 
Werkzeug==2.2.3
flask_jwt_extended==4.4.4
psycopg2-binary>=2.9
mysql-connector-python>=8.0.33
flasgger==0.9.5
requests>=2.28.0
# Para emails HTML mais bonitos (opcional)
# jinja2>=3.0.0